rename: 레이아웃 컴포넌트 분류
